requirements.txt
astroid==2.15.5
bcrypt==4.0.1
bottle==0.12.25
certifi==2023.5.7
charset-normalizer==3.1.0
dill==0.3.6
idna==3.4
isort==5.12.0
lazy-object-proxy==1.9.0
mccabe==0.7.0
platformdirs==3.5.1
PyJWT==2.7.0
pylint==2.17.4
PyMySQL==1.0.3
requests==2.30.0
tomli==2.0.1
tomlkit==0.11.8
typing_extensions==4.6.2
urllib3==2.0.2
wrapt==1.15.0
